task: Split out rating classes into separate files.
