+++ src/main/kotlin/tf/jacobsc/utils/NewPlayerRating.kt
@@ -1,0 +1,9 @@
+package tf.jacobsc.utils
+
+import de.gesundkrank.jskills.Rating
+import org.opencraft.server.model.Player
+
+data class NewPlayerRating(
+    val player: Player,
+    val rating: Rating,
+)
